add a full config strategy to the import logic so it doesn't need to always rely on the sparse config, bypassing the lean-resolution issue
